// spotify-app/spotify_app/src/interfaces/IArtists.ts
export interface ISpotifyArtists {
    items:    Item[];
    total:    number;
    limit:    number;
    offset:   number;
    href:     string;
    next:     string;
    previous: null;
}

export interface Item {
    external_urls: ExternalUrls;
    followers:     Followers;
    genres:        string[];
    href:          string;
    id:            string;
    images:        Image[];
    name:          string;
    popularity:    number;
    type:          any;
    uri:           string;
}

export interface ExternalUrls {
    spotify: string;
}

export interface Followers {
    href:  null;
    total: number;
}

export interface Image {
    height: number;
    url:    string;
    width:  number;
}